# [awsemfexporter] Restructure Metric Translator Logic