--- a/exporter/awsemfexporter/metric_translator_test.go
+++ b/exporter/awsemfexporter/metric_translator_test.go
@@ -756,7 +756,7 @@ func TestBuildCWMetric(t *testing.T) {
 	instrLibName := "InstrLibName"
 	OTelLib := "OTelLib"
 	metricSlice := []map[string]string{
-		map[string]string{
+		{
 			"Name": "foo",
 			"Unit": "",
 		},